Sticky Note v1
most of the things with the sticky note are working, few issues here and there.

when you select a sticky note, if you click and drag on the canvas it draws lines, then on click release it puts the sticky note there.

css for the pallete looks a bit off as well.