Pretty smooth pan and zoom
Fixed a few key things, sticky note coords not written correctly -> causing the sticky note to not place.

Stale reference to the canvas was causing the movement to jump back to an old point in time when you try to do a second pan -> used the updated reference to fix it.